test: refactored seeder for the job category and unit tests
