+++ src/modules/job-category/seeders/job-category.sample.ts
@@ -1,0 +1,38 @@
+const jobCategories = [
+    {
+      name: 'Front-end',
+      image_url: 'https://i.imgur.com/a2FQ9F6.png',
+    },
+    {
+      name: 'Back-end',
+      image_url: 'https://i.imgur.com/462x9OZ.png',
+    },
+    {
+      name: 'Full-stack',
+      image_url: 'https://i.imgur.com/QPOHLGX.png',
+    },
+    {
+      name: 'DevOps',
+      image_url: 'https://i.imgur.com/KRz7nkj.png',
+    },
+    {
+      name: 'Mobile',
+      image_url: 'https://i.imgur.com/dVP5a6E.png',
+    },
+    {
+      name: 'UI/UX',
+      image_url: 'https://i.imgur.com/T2T0BG0.png',
+    },
+    {
+      name: 'QA',
+      image_url: 'https://i.imgur.com/b3ykh2k.png',
+    },
+    {
+      name: 'Testers',
+      image_url: 'https://i.imgur.com/2IoUOot.png',
+    },
+  ];
+
+export default jobCategories
+
+export const TOTAL_OF_JOB_CATEGORIES: number = jobCategories.length
